requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
beautifulsoup4==4.10.0
bs4==0.0.1
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.3
Flask==2.0.3
frozenlist==1.3.0
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
jsonify==0.5
MarkupSafe==2.0.1
multidict==6.0.2
python-dotenv==0.19.2
requests==2.27.1
soupsieve==2.3.1
urllib3==1.26.8
Werkzeug==2.0.3
yarl==1.7.2
